Refactor `lakefs_spec.commithook` into generic `hooks` module
Extends the `FSEvent` enum to all currently overridden file system events.

Adds a `noop` hook that serves as the default to run in case none is
registered.
